+++ src/main/java/ch/uzh/ifi/access/course/service/StudentSubmissionService.java
@@ -1,9 +1,9 @@
 package ch.uzh.ifi.access.course.service;
 
+import ch.uzh.ifi.access.course.dao.StudentSubmissionRepository;
 import ch.uzh.ifi.access.course.model.Assignment;
 import ch.uzh.ifi.access.course.model.Exercise;
 import ch.uzh.ifi.access.course.model.workspace.StudentSubmission;
-import ch.uzh.ifi.access.course.model.workspace.StudentSubmissionRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
